+++ src/lib.rs
@@ -1,14 +1,12 @@
 // A lot of these things would eventually go into some kind of "internal" module, since
 // we don't actually want to expose them in the public API?
-mod apply;
-pub mod bdd;
 pub mod bdd_node;
 pub mod boolean_operators;
 mod conversion;
-mod nested_apply;
 pub mod node_id;
 pub mod node_table;
 pub mod shared;
+pub mod split;
 pub mod task_cache;
 pub mod variable_id;
